Breaking down the question of gamma parameter usage in IISPH 3D
After consideration and a discussion with a professor of mine we determined the following: The gamma parameter that was there previous to the single layer boundary concept is (eventually) not required. In the new single layer boundary case the parameter can be considered as gamma_2. Gamma_1 is without doubt correctly used in this case to scale the contribution of boundary particles to the density term.

We now need to check:
* If gamma can be omitted in multi layer boundary scenarios
* If single layer boundary scenarios are handled correctly