Restructured the whole sensor system
The sensor system is now more expressive and easier to use. We use json based output in order to allow for arbitrary data of the sensors in each timestep.
The new format of the sensor files are just multiple json literals following each other separated by a newline. The literals are either an object or an array. The object describes the following sensor data. The array contains as first entry an object describing the timepoint. The following entries are sensor data. Those sensor data entries were previously described by the definitions inside the object.